Move Local to Kernel.*
