+++ gtsam/geometry/Kernel.h
@@ -21,12 +21,63 @@
 #include <gtsam/base/Lie.h>
 #include <gtsam/base/Matrix.h>
 #include <gtsam/dllexport.h>
+#include <optional>
 
 namespace gtsam {
 namespace so3 {
 
-// Forward declare Local
-struct Local;
+struct Kernel;        // forward declare
+struct InvJKernel;    // forward declare
+
+/**
+ * Opaque evaluation context at ω: caches W, WW, θ, θ², nearZero/nearPi,
+ * Lazily computes D, G, dB, dC, dG on demand.
+ * Math is based on Ethan Eade's elegant Lie group document, at
+ * https://www.ethaneade.org/lie.pdf, and the Kernel idea in doc/Jacobians.md
+ */
+struct GTSAM_EXPORT Local {
+  /// Tolerance for near zero (theta^2)
+  static constexpr double kNearZeroThresholdSq = 1e-6;
+  /// Tolerance for near pi (delta^2 = (pi - theta)^2)
+  static constexpr double kNearPiThresholdSq = 1e-6;
+
+  Vector3 omega;         ///< The rotation vector.
+  double theta2;         ///< The squared norm of the rotation vector (theta^2).
+  double theta;          ///< The norm of the rotation vector (theta).
+  Matrix3 W;             ///< The skew-symmetric matrix for the rotation vector.
+  Matrix3 WW;            ///< The square of the skew-symmetric matrix (W * W).
+  bool nearZero{false};  ///< Flag indicating if theta is near zero.
+  bool nearPi{false};    ///< Flag indicating if theta is near pi.
+  double A, B, C;        ///< Ethan's A,B,C coefficients
+
+  /// Constructor with element of Lie algebra so(3)
+  explicit Local(const Vector3& omega,
+                 double nearZeroThresholdSq = kNearZeroThresholdSq,
+                 double nearPiThresholdSq = kNearPiThresholdSq);
+
+  // Exponential map via Rodrigues formula: I + A(θ) W + B(θ) WW
+  Matrix3 expmap() const;
+
+  // Jacobian kernel J_[l/r] = I +/0 B W + C WW  (left/right).
+  Kernel Jacobian() const &;
+
+  // Specialized kernel for inverse Jacobian, stable even for |omega| > π
+  InvJKernel InvJacobian() const &;  // I +/- 1/2 W + D WW
+
+  // Gamma kernel: Γ_[l/r] = 0.5 I ± C W + G WW (left/right).
+  Kernel Gamma() const &;
+
+  // access to (lazily evaluated) coefficients
+  double D() const;
+  double E() const;
+  double dB() const;
+  double dC() const;
+  double dE() const;
+
+ protected:
+  mutable std::optional<double> D_, E_;         ///< D-E are lazily computed.
+  mutable std::optional<double> dB_, dC_, dE_;  ///< Radial derivatives c(θ)'/θ
+};
 
 /**
  * Kernel: M(ω) = a I + b W + c W^2 with radial derivatives db,dc for Fréchet.
@@ -72,6 +123,32 @@ Kernel axpy(double alpha, const Kernel& X, const Kernel& Y);
 
 // Blend K = α X + (1-α) Y with radial derivative (·)'/θ via dalpha
 Kernel blend(double alpha, double dalpha, const Kernel& X, const Kernel& Y);
+
+/// @deprecated: use so3::Local
+struct GTSAM_EXPORT ExpmapFunctor : public Local {
+  explicit ExpmapFunctor(const Vector3& omega);
+  ExpmapFunctor(double nearZeroThresholdSq, const Vector3& axis);
+  ExpmapFunctor(const Vector3& axis, double angle);
+};
+
+/// @deprecated: use so3::Local
+struct GTSAM_EXPORT DexpFunctor : public ExpmapFunctor {
+  using J33 = OptionalJacobian<3, 3>;
+  explicit DexpFunctor(const Vector3& omega, double nearZeroThresholdSq = 1e-6,
+                       double nearPiThresholdSq = 1e-6);
+  Matrix3 rightJacobian() const;
+  Matrix3 leftJacobian() const;
+  Matrix3 rightJacobianInverse() const;
+  Matrix3 leftJacobianInverse() const;
+  Vector3 applyRightJacobian(const Vector3& v, J33 H1 = {}, J33 H2 = {}) const;
+  Vector3 applyRightJacobianInverse(const Vector3& v, J33 H1 = {},
+                                    J33 H2 = {}) const;
+  Vector3 applyLeftJacobian(const Vector3& v, J33 H1 = {}, J33 H2 = {}) const;
+  Vector3 applyLeftJacobianInverse(const Vector3& v, J33 H1 = {},
+                                   J33 H2 = {}) const;
+  inline Matrix3 dexp() const { return rightJacobian(); }
+  inline Matrix3 invDexp() const { return rightJacobianInverse(); }
+};
 
 }  // namespace so3
 }  // namespace gtsam
